#43 move components to /components
