+++ app/list/layout.tsx
@@ -2,7 +2,7 @@ import '@/app/global.css'
 
 import styles from "./styles.module.css";
 import { draftMode } from "next/headers";
-import AlertBanner from "../(home)/alert-banner";
+import AlertBanner from "../../components/alert-banner";
 import { VisualEditing } from "next-sanity";
 import { SpeedInsights } from "@vercel/speed-insights/next";
